@@ -5,20 +5,24 @@
   "main": "index.js",
   "private": true,
   "dependencies": {
+    "history": "^1.13.1",
     "keymaster": "^1.6.2",
     "lodash": "^3.2.0",
-    "react": "^0.13.3",
-    "react-router": "^0.13.3",
+    "react": "^0.14.3",
+    "react-addons-linked-state-mixin": "^0.14.3",
+    "react-dom": "^0.14.3",
+    "react-router": "^1.0.2",
     "reflux": "^0.2.5",
     "reqwest": "^1.1.5",
     "store": "^1.3.17"
   },
   "devDependencies": {
-    "babelify": "^6.3.0",
-    "browserify": "^11.1.0",
-    "reactify": "^1.1.0",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babelify": "^7.2.0",
+    "browserify": "^12.0.0",
     "uglifyify": "^3.0.1",
-    "watchify": "^2.3.0"
+    "watchify": "^3.6.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 0"
