@@ -1,13 +1,15 @@
 var isElementInViewport = require("../../util/elementscrolling.js").isElementInViewport;
 var scrollIntoView = require("../../util/elementscrolling.js").scrollIntoView;
-var React = require("react/addons");
+var React = require("react");
+var ReactDOM = require('react-dom');
+
 
 module.exports = React.createClass({
     componentDidMount: function() {
       if(!this.state.hasAnchored) {
         var anchor = this.props.parent.refs.anchor;
         if(anchor) {
-          var element = React.findDOMNode(anchor);
+          var element = ReactDOM.findDOMNode(anchor);
           if(element) {
             if(!isElementInViewport(element)) {
               scrollIntoView(element);
