@@ -1,5 +1,5 @@
 var _ = require("lodash");
-var React = require("react/addons");
+var React = require("react");
 var PostLink = require("./postlink.js");
 
 
